+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.3
+
+### Changed
+
+- Add cache_service for get expiry cache url
+- Fix error with rspec gem versions
+- Add crawler_service to warmup urls
+
 ## 0.1.2
 
 ### Changed